Add router-based top navigation
