+++ package.json
@@ -12,7 +12,8 @@
   "dependencies": {
     "canvas-confetti": "^1.9.3",
     "react": "^19.1.0",
-    "react-dom": "^19.1.0"
+    "react-dom": "^19.1.0",
+    "react-router-dom": "^6.22.3"
   },
   "devDependencies": {
     "@eslint/js": "^9.29.0",
