+++ src/pages/Minigames.jsx
@@ -1,0 +1,14 @@
+import Layout from '../components/Layout';
+
+function Minigames() {
+  return (
+    <Layout>
+      <div className="text-center text-green-400 mt-10">
+        <h2 className="text-3xl mb-4">Minigames</h2>
+        <p>More nostalgic games will appear here soon.</p>
+      </div>
+    </Layout>
+  );
+}
+
+export default Minigames;
