+++ src/pages/Profile.jsx
@@ -1,0 +1,14 @@
+import Layout from '../components/Layout';
+
+function Profile() {
+  return (
+    <Layout>
+      <div className="text-center text-green-400 mt-10">
+        <h2 className="text-3xl mb-4">Profile</h2>
+        <p>Your profile details will live here.</p>
+      </div>
+    </Layout>
+  );
+}
+
+export default Profile;
